+++ CHANGELOG.md
@@ -1,10 +1,15 @@
 ## [0.0.1] -  Initial Release.
 
-## [0.1.0] -  Stable Release 
+## [0.1.0] -  Stable Release : 
 Added Support for listening the updates via onUpdate method. It can be used for pushing routes or showing snackbar. See More in Example
 
-## [0.1.2] -  New Function 
+## [0.1.2] -  New Function :
 Added Support for relaying more than one updates at a time via relayMultiple() function.
 
-## [0.1.4] -  Major Bug Fixes;
+## [0.1.4] -  Major Bug Fixes:
 Bug of re-initialization of station object fixed.
+
+## [0.2.0] - Major Api Breaking Changes :
+- Renamed Station to Store.
+- Provider Widgets are decoupled using ProviderMixin.
+- Store Initialization made global and uplifted to Provider Widget.